==> R/00_all.qmd <==
# Master document

01_load.qmd contains the script that loads all the raw data we used in the project

{{< include 01_load.qmd >}}

02_clean.qmd contains the cleaning process we followed to optimize the data format for the project

{{< include 02_clean.qmd >}}

## Creation of new variables

This file contains the new variables useful for our analysis

{{< include 03_augment.qmd >}}

## Analysis

These two files contain the data analysis including plots and figures.

{{< include 05_analysis_1.qmd >}}

{{< include 06_analysis_2.qmd >}}

---
title: "00_all.qmd"
author: "group_12"
format: 
  html:
    embed-resources: true
editor: visual
---
