==> R/01_load.qmd <==
---
title: "01_load.qmd"
author: "group_12"
format: 
  html:
    embed-resources: true
editor: visual
---

## Extract data

Load all the datasets we used and processed throughout the study analysis

```{r}
#| echo: true
#| message: false
#| warning: false
#| label: setup
library("tidyverse")
library("maps")
library("ggplot2")
library("viridis")

```

**Load data**

```{r}
#| echo: true
#| eval: true
filename <-  c("DONsRaw.csv",
               "COVIDOutbreaks.csv",
               "Outbreaks.csv",
               "UniqueDONs.csv",
               "icd1011.csv",
               "isocodes.csv",
               "WDICountry.csv", 
               "world-administrative-boundaries.csv")
dir_name <- "../_raw/"

if( !dir.exists(dir_name) ){
  dir.create(path = dir_name)
}

if ( length(list.files(dir_name, 
                       pattern="csv")) != length(filename) ) {
  print(str_c(str_flatten_comma(filename,), 
              " required in _raw folder"))

} else if ( length(list.files(dir_name, 
                              pattern="csv")) == length(filename)) { 
  rawdons <-  read.csv(str_c(dir_name,
                             "DONsRaw.csv"))
  covidOutbreaks <- read.csv(str_c(dir_name,
                                   "COVIDOutbreaks.csv"), 
                             sep = ",")
  outbreaks <- read.csv(str_c(dir_name,
                              "Outbreaks.csv"), 
                        sep = ",")
  uniqueDONs <- read.csv(str_c(dir_name,
                               "UniqueDONs.csv"),
                         sep = ",")
  wdi <- read.csv(str_c(dir_name,"WDICountry.csv"))
  worldboundaries <- read.csv(str_c(dir_name, 
                                    "world-administrative-boundaries.csv"), 
                              sep = ";")
  icd1011 <- read.csv(str_c(dir_name,
                            "icd1011.csv"), 
                      sep =";")
  isocodes <- read.csv(str_c(dir_name,
                             "isocodes.csv"), 
                       sep = ";")
}
```


Writing TSV
```{r}
write_tsv(rawdons, "../Data/rawdons.tsv")
write_tsv(covidOutbreaks, "../Data/covidOutbreaks.tsv")
write_tsv(outbreaks, "../Data/outbreaks.tsv")
write_tsv(uniqueDONs, "../Data/uniqueDONs.tsv")
write_tsv(wdi, "../Data/wdi.tsv")
write_tsv(worldboundaries, "../Data/worldboundaries.tsv")
write_tsv(icd1011, "../Data/icd1011.tsv")
write_tsv(isocodes, "../Data/isocodes.tsv")
```
