==> R/02_clean.qmd <==
---
title: "02_clean.qmd"
author: "group_12"
format: 
  html:
    embed-resources: true
editor: visual
---


```{r}
library("tidyverse")
library("readr")

#| echo: true
#| message: false
#| warning: false
#| eval: false
```

## LOADING DATA

```{r}
rawdons <- read_tsv("../Data/rawdons.tsv")
covidOutbreaks <- read_tsv("../Data/covidOutbreaks.tsv")
outbreaks <- read_tsv( "../Data/outbreaks.tsv")
uniqueDONs <- read_tsv( "../Data/uniqueDONs.tsv")
wdi <- read_tsv( "../Data/wdi.tsv")
worldboundaries <- read_tsv( "../Data/worldboundaries.tsv")
icd1011 <- read_tsv( "../Data/icd1011.tsv")
isocodes <- read_tsv( "../Data/isocodes.tsv")

```


## CLEANING DATA

This file contains the steps we followed to clean the data-sets we used in the study analysis.



## CLEAN WDI DATA:

**We keep only the variables needed, and rename Country.Code to iso3 for merging purposes. Renaming Kosovo iso3, as there is no universally accepted code, and they differ in the world bank (XKX) and WHO (XXK). We filter to include only countries with income data available. :**

```{r}
wdi_clean <- wdi |> 
  select(Country.Code, Income.Group) |> 
  rename(iso3 = Country.Code) |> 
  mutate(iso3 = if_else(iso3 == "XKX", 
                        "XXK", 
                        iso3)) |> 
  filter(Income.Group !="")

write_tsv(wdi_clean, "../Data/wdi_clean.tsv")

```

## CLEAN WORLD BOUNDARIES DATA:

**Removing rows with blank cells:**

```{r}
worldboundaries <- worldboundaries |> 
  filter_all(all_vars(. != ""))
```

**Renaming iso3, Continent and Coordinates column and removing not needed columns:**

```{r}
worldboundaries <- worldboundaries |> 
  rename(
    iso3 = "ISO.3.country.code",
    Continent = "Continent.of.the.territory",
    Coordinates = "Geo.Point") |> 
  select(-French.Name, 
         -English.Name,
         -ISO.3.territory.code, 
         -Region.of.the.territory, 
         -ISO.3166.1.Alpha.2.Codes, 
         -Geo.Shape)
```

**Correcting iso3 and discarding rows with wrong iso3-Continent combination:**

```{r}
worldboundaries <- worldboundaries |>
  mutate(
    iso3 = case_when(
      iso3 == "United States of America" ~ "USA",
      iso3 == "Brazil" ~ "BRA",
      TRUE ~ iso3
    ),
    Continent = case_when(
      Continent %in% c("Northern America", 
                       "South America") ~ "Americas",
      TRUE ~ Continent
    )
  ) |>
  filter(
    !(Continent != "Americas" & iso3 == "USA") &
    !(Continent != "Europe" & iso3 %in% c("GBR", 
                                          "FRA"))
  )
```

**Removing rows with duplicates of the same iso3:**

```{r}
worldboundaries <- worldboundaries |>
  distinct(iso3, 
           .keep_all = TRUE)
```

## CLEAN RAW DONs:

**Changing the column names:**

```{r}
rawdons_clean <- rawdons|>   
  rename(DONs = ID) |>    
  mutate(Year = as.integer(str_split_i(Date, 
                                       pattern = "-", 
                                       1)))    
write_tsv(rawdons_clean, "../Data/rawdons_clean.tsv")
```

## CLEAN UNIQUE DONs and COVID-19 DATA:

**Rename of column, handling null and unspecific values:**

```{r}
uniqueDONs_clean <- uniqueDONs |>
  rename(Index = X) |>
  mutate(Definition = if_else(Definition == "", 
                              "Unspecified", 
                              Definition),
         icd11l2 = if_else(icd11l2 == "", 
                           "Not reported", 
                           icd11l2),
                           icd11l3 = if_else(icd11l3 == "", 
                                             "Not reported", 
                                             icd11l3),
                           icd11l2 = if_else(str_detect(icd11l2, 
                                                        "Aetiology"), 
                                             "Bacteria or viral infection", 
                                             icd11l2))
  
covidOutbreaks_clean <- covidOutbreaks |>
  rename(Index = X) |>
  mutate(
    icd11c1 = as.character(icd11c1),
    icd11l2 = if_else(str_detect(icd11l2, 
                                 "International"),
                      "New uncertain disease", 
                      icd11l2))

write_tsv(uniqueDONs_clean, "../Data/uniqueDons_clean.tsv")
write_tsv(covidOutbreaks_clean, "../Data/covidOutbreaks_clean.tsv")
```

## MERGING DATA

**Merging of unique and COVID-19 outbreaks data:**

```{r}
total_outbreaks <- 
  full_join(uniqueDONs_clean, 
            covidOutbreaks_clean) |>
  distinct()
```

**Merging of WDI and total outbreaks data:**

```{r}
outbreaks_wdi <-  
  left_join(total_outbreaks, 
            wdi_clean, 
            by = "iso3")
```

**Merging of geographical and total outbreaks data:**

```{r}
clean_outbreaks <- worldboundaries |> 
  left_join(drop_na(outbreaks_wdi), 
            by = "iso3")
```

## Cleaning merged data

**Correcting/Shortening countries names:**

```{r}
clean_outbreaks <- clean_outbreaks |> 
  mutate(
    Country = case_when(
      Country == "Congo Democratic Republic of the" ~ "Democratic Republic of Congo",
      Country == "United States of America" ~ "USA",
      Country== "C̫te d'Ivoire" ~ "Côte d'Ivoire",
      Country == "United Kingdom of Great Britain and Northern Ireland" ~ "United Kingdom",
      TRUE ~ as.character(Country)
    )
  )
```

**Creating a final clean data file: clean_outbreaks**

```{r}
write_tsv(clean_outbreaks, "../Data/clean_outbreaks.tsv")
```
