==> R/06_analysis_2.qmd <==
---
title: "06_analysis_2.qmd"
author: "group_12"
format:
  html:
    embed-resources: true
editor: visual
---

```{r}

library("tidyverse")
library("ggplot2")

#| echo: true
#| message: false
#| warning: false
#| eval: false
```

## Including the final clean and augmented dataset

```{r}
final_outbreaks <- read_tsv("../Data/final_outbreaks.tsv")
```

## Disease-outbreaks, reproducing fig4b with top 20 diseases

```{r}
#| echo: true
#| eval: false
#keeping the 20 most frequent diseases
#the fct_infreq orders the Disease in levels according to frequency
#the str_wraps wraps text that is over 30 characters

Fig4b_reproduced <- final_outbreaks |> 
  count(Disease) |> 
  arrange(desc(n)) |> 
  slice_head(n = 20) |>  
  inner_join(final_outbreaks, by = "Disease")  |> 
  mutate(Disease = fct_infreq(Disease))  |>  
  ggplot(mapping = aes(x = Disease)) +
  geom_bar()+
  labs(y = "Total frequency of outbreaks \n (1996 - 2021)", 
       x = "Disease") +
  scale_x_discrete(labels = function(x) str_wrap(x, 
                                                 width = 30)) + 
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 90, 
                                   vjust = 0.5,
                                   hjust = 1))

Fig4b_reproduced

ggsave(Fig4b_reproduced, 
       filename = "../Results/Fig4b_reproduced.png", 
       dpi = 300)
```

## Creating a new Figure 4b, with incorporated information on country income status

```{r}

Fig4b_income <- final_outbreaks |> 
  count(Disease) |> 
  arrange(desc(n)) |> 
  slice_head(n = 20) |> 
  inner_join(final_outbreaks, 
             by = "Disease")  |> 
  mutate(Disease = fct_infreq(Disease), 
         Income.Group = factor(Income.Group, 
                                levels = c("High income", 
                                           "Upper middle income", 
                                           "Lower middle income", 
                                           "Low income", 
                                           "No income data available"))) |>
  ggplot(mapping = aes(x = Disease, 
                       fill = Income.Group)) +
  geom_bar()+
  scale_fill_brewer(palette = "Spectral") +
  labs(y = "Total frequency of outbreaks \n (1996 - 2021)", 
       x = "Disease") +
  scale_x_discrete(labels = function(x) str_wrap(x, 
                                                 width = 30)) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 90, 
                                   vjust = 0.5, 
                                   hjust = 1, 
                                   size = 8), 
        legend.title = element_blank())

Fig4b_income

ggsave(Fig4b_income, 
       filename = "../Results/Fig4b_income.png",
       dpi = 300)

```

##descriptive data of fig 4b income

```{r}

#distribution of countries by income status in the entire dataset (almost equal - between 20 and 30 %). 
income_dist <- final_outbreaks |> 
  group_by(Income.Group) |> 
  summarise(count = n())  |> 
  mutate(percentage = (count / sum(count)) * 100)
income_dist

#distribution of countries by income status for the top 20 diseases (still almost equal)
income_dist_top20 <- final_outbreaks |> 
count(Disease) |> 
  arrange(desc(n)) |> 
  slice_head(n = 20) |>  
  inner_join(final_outbreaks, by = "Disease")  |> 
  group_by(Income.Group) |> 
  summarise(count = n())  |> 
  mutate(percentage = (count / sum(count)) * 100)
income_dist_top20


#Now excluding covid and influenza, and looking at the distribution of countries by income status for the top 3-20 diseases (So here, the disease burden changes with 66% of the outbreaks in Low income or lower middle income countries.)
income_dist_top3_20 <- final_outbreaks |> 
filter(Disease != "COVID-19", 
       Disease !="Influenza due to identified zoonotic or pandemic influenza virus") |> 
  count(Disease) |> 
  arrange(desc(n)) |> 
  slice_head(n = 18) |>  
  inner_join(final_outbreaks, by = "Disease")  |> 
  group_by(Income.Group) |> 
  summarise(count = n())  |> 
  mutate(percentage = (count / sum(count)) * 100)
income_dist_top3_20

```

## Disease type distribution over the years 1996-2022

```{r}
Disease_distribution <- final_outbreaks |>
  drop_na() |> 
  ggplot(mapping =
           aes(x = Year,
               fill = Disease_class, 
               group = Disease_class)) +
  geom_bar(stat = "count") +
  scale_x_continuous(breaks = seq(1996, 
                                  2022, 
                                  by = 1)) +
  scale_y_continuous(breaks = seq(min(0), 
                                  max(300), 
                                  by = 20)) +  
  theme_minimal() +
  scale_fill_viridis_d(alpha = 0.8) +
  theme(plot.title = element_text(size = 12,
                                  vjust = -2,
                                  hjust = 0.5),
        axis.title.x = element_text(size = 7),
        axis.title.y = element_text(size = 8),
        axis.text.x = element_text(size = 6,
                                   angle = 45),
        axis.text.y = element_text(size = 6),
        legend.title = element_text(size = 8),
        legend.text = element_text(size = 6)) +
  labs(x = "Year",
       y = "Count of disease",
       title = "Disease type distribution over time",
       fill = "Disease classification")

 ggsave(Disease_distribution, 
        filename = "../Results/Disease_distribution.png", 
        dpi = 300)
Disease_distribution
```

## Frequency of outbreaks for countries grouped by income in years 1996-2022

```{r}
# calculating the number of outbreaks in countries grouped by income
A1 <- final_outbreaks |> 
  select(Income.Group,
         Year) |> 
  drop_na() |> 
  mutate(one = 1) |> 
  pivot_wider(names_from =Income.Group,
              values_from = one, 
              values_fn = sum)

A2 <- A1 |> 
  pivot_longer(cols = -Year, 
               names_to ="Income.Group",
               values_to = "Frequency")  


#arranging data in an appropriate order
A2 <- A2 |> 
  mutate(Income.Group = factor(Income.Group, 
                          levels=c("High income",
                                   "Upper middle income",
                                   "Lower middle income",
                                   "Low income"
                                   )))

#making plot
Inc_line <- ggplot(data = A2, 
                aes(x = Year, 
                    y = Frequency,
                    color = Income.Group,
                    group = Income.Group
                    )) + 
          geom_line(size = 1) +
      ylab("Frequency of outbreaks") + 
      xlab("Year") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45)) +
  theme(legend.position = "right")+
   #scale_color_viridis_d() +
  scale_x_continuous(breaks = seq(1996,
                                  2022, 
                                  by = 2))+
   labs(color = "Income group")

#saving graph
ggsave(Inc_line, 
       filename ="../Results/Inc_line.png", 
       dpi = 300)

Inc_line
```

## Frequency of outbreaks for each continent in years 1996-2022

```{r}
# calculating the number of outbreaks in countries grouped by continent
B1 <- final_outbreaks |>
  select(Continent,
         Year) |> 
  mutate(one = 1) |>
  pivot_wider(names_from = Continent,
              values_from = one, 
              values_fn = sum)
B2 <- B1 |> 
  pivot_longer(cols = -Year, 
               names_to = "Continent",
               values_to = "Frequency") 
#replacing NA's with 0 value as there was no information of an outbreak for a specific continent in a specific year
B2 <-B2 |> 
  mutate(Frequency = replace_na(Frequency, 0))

#plot Frequency of outbreaks for each continent in years 1996-2022
Cont_line <- ggplot(data = B2, 
                aes(x = Year, 
                    y = Frequency,
                    color = Continent,
                    group = Continent)) +  
          geom_line(size = 1) +
      ylab("Frequency of outbreaks") + 
      xlab("Year") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45)) +
  theme(legend.position="right")+
  scale_color_viridis_d() +
  scale_x_continuous(breaks = seq(1996,
                                  2022, 
                                  by = 2))
#saving graph
ggsave(Cont_line, 
       filename = "../Results/Cont_line.png", 
       dpi = 300)

Cont_line
```
